requirements.txt
streamlit>=1.29.0
google-cloud-aiplatform[adk,agent_engines]>=1.55.0
langchain>=0.1.0
langchain-google-vertexai>=0.1.0
langchain-community>=0.0.13
pydantic>=2.0.0
python-dotenv>=1.0.0
faiss-cpu>=1.7.4
pylint>=3.0.0
python-magic>=0.4.27
unstructured>=0.11.0
pdf2image>=1.16.3
python-docx>=1.0.1
openpyxl>=3.1.2
bs4>=0.0.1
pytesseract>=0.3.10
